feat: Enhance local configurations and ECS task setup
This commit introduces more granular secret handling for database and SQL server parameters, adjusts the ECR repository for viewer, and adds 'nbs auth' variable. In the ECS task configuration, a few more properties are specified for container definitions, and the re-deployment trigger in ECS service has been updated to include version in addition to timestamp.